(half) Bugfix #2: Make the title reflect countdown
I don't understand how this change actually does that...
still trying to work it out!

Before this change, if you start and stop the timer a bit later, the
title would be one second behind the duration.

The change in timer.coffee is just to prevent too many calls to
@update() as @reset() will result in an @update() call anyway.