Implement a "night" theme
At night, a bright white screen is not very comfortable on the eyes.
This theme makes everything dark if there is a class of "night" on <body>

Currently the UI does not provide a way to switch to this theme because
I'm not sure how to do it while keeping the interface clean.

The javascript to turn it on can be as simple as:

```
$('body').toggleClass("night")
```